feature: add layouting options to templates
