+++ example/content/content-box.md
@@ -2,7 +2,7 @@
 description: Description for Content Box page
 id: a0497490-b41a-11eb-ac27-59ada86bb225
 keywords: Test1, Test2, Test3
-layout: default
+layout: content
 slug: content-box
 title: Content Box
 sections:
